requirements.txt
fastapi
uvicorn
requests
newspaper3k
trafilatura
python-dotenv
google-generativeai